"添加水鱼查分器排名API支持及隐私设置处理"
此次代码变更主要新增了水鱼查分器排名API的支持以及对用户隐私设置的处理。具体变更内容包括：

1. 新增了`DivingFishRatingRankApi`类，用于处理水鱼查分器的排名数据。这包括数据的初始化、更新和查询，以及通过用户名查找对应的Rating和排名。

2. 在`app.py`文件中增加了对图片隐私设置的处理逻辑。如果用户设置了隐私保护，将生成一个隐私文件并返回相关提示信息，以保护用户数据不被直接显示。

3. 修改了`generate_img.py`文件中生成图片的逻辑，以支持隐私设置。当用户生成B50图片时，如果生成失败，将创建一个隐私文件，防止因隐私设置或其他原因导致的图片生成失败。

4. 在`text_chat_handler.py`中添加了对新API的支持，包括处理水鱼查分器排名查询的命令`我有多菜`。这允许用户通过该命令查询自己的Rating和排名信息。

此次变更主要围绕着增强用户体验和保护用户隐私进行，确保用户能够在享受服务的同时，拥有更好的数据安全保障。

+++ maibox/generate_img.py
@@ -604,12 +604,19 @@ def draw_text_with_spacing(draw: ImageDraw.ImageDraw, pos, text, font, fill_colo
 def call_b50(fish_username, filename, nickname=None, icon_id=None, wechat_utils: WechatInterface = None, non_hashed_wxid: str=""):
     with open(f"img/{filename}.flag", "wb") as f:
         f.write(b"")
-    B50Img: Image = generate({'username': fish_username, 'b50': True}, nickname, icon_id, filename).convert("RGB")
-    B50Img.save(f"./img/{filename}", format="png", quality=90)
+    B50Img = generate({'username': fish_username, 'b50': True}, nickname, icon_id, filename)
+    if B50Img:
+        B50Img.save(f"./img/{filename}", format="png", quality=90)
+    else:
+        with open(f"img/{filename}.privacy", "w") as f:
+            f.write("")
 
     if wechat_utils and wechat_utils.interface_test():
-        wechat_utils.send_image(f"./img/{filename}", non_hashed_wxid)
-        os.remove(f"./img/{filename}")
+        if B50Img:
+            wechat_utils.send_image(f"./img/{filename}", non_hashed_wxid)
+            os.remove(f"./img/{filename}")
+        else:
+            wechat_utils.send_text("由于用户隐私设置或其他水鱼服务器原因，图片生成失败！", non_hashed_wxid)
 
     os.remove(f"./img/{filename}.flag")
